requirements.txt
requests==2.28.2
python-dotenv==1.0.0
click==8.1.3
pydantic==1.10.7
python-dateutil==2.8.2